requirements.txt
aenum==3.1.16
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
billiard==4.2.1
boto3==1.37.3           # brings in botocore 1.37.3
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.3
dataclasses-json==0.6.7
deepgram-sdk==4.1.0
deprecation==2.1.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
kombu==5.5.3
lxml==5.4.0
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
packaging==25.0
pdfminer==20191125
pdfminer.six==20250506
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-magic==0.4.27
python-multipart==0.0.20
redis==6.0.0
s3transfer>=0.11.0,<0.12.0  # satisfies boto3 1.37.3’s requirement
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
pyyaml